feature(server): create more routes and make the secure
- create routes to cover specifications in the rubrics file
- integrate user sign up and login with jwt
- refactor server route
- implement feedabck on UI template
- add thumbnails and auto fill for longitude and latitude

[Delivers #162531703 #162472056 #162472163 #162617061]

+++ src/controllers/records.js
@@ -1,6 +1,5 @@
 /* eslint-disable no-console */
 /* eslint-disable import/no-unresolved */
-import moment from 'moment';
 import db from '../db';
 
 const Record = {
@@ -12,19 +11,18 @@ const Record = {
    */
   async createRecord(req, res) {
     const text = `INSERT INTO
-      records(title, description, type, location, status, comment,message, created_on, modefied_on)
-      VALUES($1, $2, $3, $4, $5, $6, $7, $8, $9)
+      records(title, description, type, owner_id, location, status, comment, message)
+      VALUES($1, $2, $3, $4, $5, $6, $7, $8)
       returning *`;
     const values = [
       req.body.title,
       req.body.description,
       req.body.type,
+      req.user.id,
       req.body.location,
       'draft',
       req.body.comment,
       'record created',
-      moment(new Date()),
-      moment(new Date()),
     ];
 
     try {
@@ -34,6 +32,7 @@ const Record = {
         data: rows[0],
       });
     } catch (error) {
+      console.log(error.message);
       return res.status(400).send({
         status: 400,
         message: 'enter a valid type e.g (red-flag, intervention)',
@@ -58,7 +57,7 @@ const Record = {
     } catch (error) {
       return res.status(500).send({
         status: 500,
-        message: 'server error',
+        message: error.message,
       });
     }
   },
@@ -66,12 +65,12 @@ const Record = {
    * Get A Record
    * @param {object} req
    * @param {object} res
-   * @returns {object} reflection object
+   * @returns {object} record object
    */
   async getOneRecord(req, res) {
-    const text = 'SELECT * FROM records WHERE id = $1';
+    const text = 'SELECT * FROM records WHERE id = $1 AND owner_id = $2';
     try {
-      const { rows } = await db.query(text, [req.params.id]);
+      const { rows } = await db.query(text, [req.params.id, req.user.id]);
       if (!rows[0]) {
         return res.status(404).send({
           status: 404,
@@ -95,43 +94,47 @@ const Record = {
    * @param {object} res
    * @returns {object} updated record
    */
-  async updateLocation(req, res) {
-    const findOneQuery = 'SELECT * FROM records WHERE id=$1';
-    const updateOneQuery = `UPDATE records
-      SET location=$1,modefied_on=$2
-      WHERE id=$3 returning *`;
+  async findByType(req, res) {
+    const findOneQuery = 'SELECT * FROM records WHERE type = $1';
     try {
-      const { rows } = await db.query(findOneQuery, [req.params.id]);
-      if (!rows[0]) {
+      const { rows } = await db.query(findOneQuery, [req.params.type]);
+      if (!rows) {
         return res.status(404).send({
           status: 404,
           message: 'record not found',
         });
       }
-      const values = [
-        req.body.location,
-        moment(new Date()),
-        req.params.id,
-      ];
-      const response = await db.query(updateOneQuery, values);
       return res.status(200).send({
         status: 200,
-        data: response.rows[0],
+        data: rows,
       });
-    } catch (err) {
+    } catch (error) {
       return res.status(400).send({
         status: 400,
-        error: 'enter a valid id',
+        error: error.message,
       });
     }
   },
-  async updateComment(req, res) {
-    const findOneQuery = 'SELECT * FROM records WHERE id=$1';
+  /**
+   * Update A Record
+   * @param {object} req
+   * @param {object} res
+   * @returns {object} updated record
+   */
+  async update(req, res) {
+    let type;
+    if (req.body.location) {
+      type = 'location';
+    } else if (req.body.comment) {
+      type = 'comment';
+    }
+
+    const findOneQuery = 'SELECT * FROM records WHERE id = $1 AND owner_id = $2';
     const updateOneQuery = `UPDATE records
-      SET comment=$1,modefied_on=$2
-      WHERE id=$3 returning *`;
+    SET ${type}=$1
+    WHERE id=$2 AND owner_id = $3 returning *`;
     try {
-      const { rows } = await db.query(findOneQuery, [req.params.id]);
+      const { rows } = await db.query(findOneQuery, [req.params.id, req.user.id]);
       if (!rows[0]) {
         return res.status(404).send({
           status: 404,
@@ -139,9 +142,9 @@ const Record = {
         });
       }
       const values = [
-        req.body.comment,
-        moment(new Date()),
+        req.body.location || req.body.comment || req.body.status,
         req.params.id,
+        req.user.id,
       ];
       const response = await db.query(updateOneQuery, values);
       return res.status(200).send({
@@ -151,7 +154,36 @@ const Record = {
     } catch (err) {
       return res.status(400).send({
         status: 400,
-        error: 'enter a valid id',
+        error: err.message,
+      });
+    }
+  },
+  /**
+   * Update A Record
+   * @param {object} req
+   * @param {object} res
+   * @returns {object} updated record
+   */
+  async updateStatus(req, res) {
+    const updateOneQuery = `UPDATE records
+    SET status=$1
+    WHERE id=$2 returning *`;
+    try {
+      const { rows } = await db.query(updateOneQuery, [req.body.status, req.params.id]);
+      if (!rows[0]) {
+        return res.status(404).send({
+          status: 404,
+          message: 'record not found',
+        });
+      }
+      return res.status(200).send({
+        status: 200,
+        data: rows[0],
+      });
+    } catch (err) {
+      return res.status(400).send({
+        status: 400,
+        error: err.message,
       });
     }
   },
@@ -159,12 +191,12 @@ const Record = {
    * Delete A Record
    * @param {object} req
    * @param {object} res
-   * @returns {void} return status code 204
+   * @returns {void} return status code 200
    */
   async delete(req, res) {
-    const deleteQuery = 'DELETE FROM records WHERE id=$1 returning *';
+    const deleteQuery = 'DELETE FROM records WHERE id=$1 AND owner_id = $2 returning *';
     try {
-      const { rows } = await db.query(deleteQuery, [req.params.id]);
+      const { rows } = await db.query(deleteQuery, [req.params.id, req.user.id]);
       if (!rows[0]) {
         return res.status(404).send({
           status: 404,
